move bkz_stats to tools
